added new text to register page to make it more clear to sign up as a user to create event. Added more text to creation start page to make it clear you need to create organization first and then event, added covid to description if they are doing non remote location, fixed ticket types issues where deleting it wasnt coming back, fixed event price range to get it based on current events (still no fix for large numbers), added a tag limit of 10 per event, added tags to each event show page and made the admin ones clickable, removed contact buttons

+++ app/Event.php
@@ -367,10 +367,12 @@ class Event extends Model
     */
     public function deleteEvent($event) 
     {
-        if ($event->moderatorcomments()->exists()) {
-            Conversation::find($event->moderatorcomments()->first()->conversation_id)->delete();
+        if ($event->user_id == auth()->id()) {
+            if ($event->conversation()->exists()) {
+                $event->conversation()->first()->delete();
+            }
+            $event->delete();
         }
-        $event->delete();
     }
 
     /**
